Added a possible common issue to README.md
Changed the game link to http:// instead of https:// (a Connect Four game really doesn't need encryption in the first place), and ddev provides access to both http:// and https:// versions.
Also added a possible solution in case the browser redirects to HTTPS and there's a warning due to a self signed certificate.